+++ src/ActionTypes.ts
@@ -1,6 +1,8 @@
-export default {
-  REQUEST: 'REQUEST',
-  ERROR: 'ERROR',
-  SUCCESS: 'SUCCESS',
-  CANCEL: 'CANCEL'
-};
+enum ActionTypes {
+  REQUEST = 'REQUEST',
+  ERROR = 'ERROR',
+  SUCCESS = 'SUCCESS',
+  CANCEL = 'CANCEL'
+}
+
+export default ActionTypes;