+++ src/ApolloNetworkStatusDispatcherContext.tsx
@@ -1,4 +1,0 @@
-import React from 'react';
-import Dispatcher from './Dispatcher';
-
-export default React.createContext<Dispatcher | undefined>(undefined);