+++ src/index.ts
@@ -1,13 +1,6 @@
-export {
-  default as ApolloNetworkStatusProvider
-} from './ApolloNetworkStatusProvider';
-export {
-  default as useApolloNetworkStatus,
-  OperationError,
-  NetworkStatus
-} from './useApolloNetworkStatus';
-export {
-  default as useApolloNetworkStatusReducer
-} from './useApolloNetworkStatusReducer';
 export {default as ActionTypes} from './ActionTypes';
-export {NetworkStatusAction} from './NetworkStatusAction';
+export {default as NetworkStatusAction} from './NetworkStatusAction';
+export {
+  default as createNetworkStatusNotifier
+} from './createNetworkStatusNotifier';
+export {OperationError, NetworkStatus} from './useApolloNetworkStatus';