+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 4.0
+
+### Improvements
+
+- Improved types for `NetworkStatusAction`.
+- A more reliable integration with the Apollo cache. This fixes [#22](https://github.com/molindo/react-apollo-network-status/issues/22) and [#28](https://github.com/molindo/react-apollo-network-status/issues/28).
+
+### Breaking changes
+
+- The usage of the library has changed from using a provider to configuring a link that needs to be passed to the `ApolloClient` constructor (see [README](./README.md)).
+- Scoping network status handling to a subtree was removed along with the `enableBubbling` option. The network status handling is scoped to the usage of the `ApolloClient` instance.
+
 ## 3.0
 
 ### New features